+++ src/nodejs/lambda-handlers/test/handlers/submission.test.js
@@ -76,7 +76,7 @@ describe('submission', () => {
         user_id: 'test user'
       }
     };
-    db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_ADMINREAD'], user_groups: [] });
+    db.user.findById.mockReturnValueOnce({ user_privileges: ['ADMIN'], user_groups: [] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_DAACREAD'], user_groups: [] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_READ'], user_groups: [] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_WRITE'], user_groups: [] });
@@ -96,7 +96,7 @@ describe('submission', () => {
         user_id: 'test user'
       }
     };
-    db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_ADMINREAD'], user_groups: [] });
+    db.user.findById.mockReturnValueOnce({ user_privileges: ['ADMIN'], user_groups: [] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_DAACREAD'], user_groups: [] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_READ'], user_groups: [] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_WRITE'], user_groups: [] });
@@ -216,7 +216,7 @@ describe('submission', () => {
       ];
     });
 
-    db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_ADMINREAD'] });
+    db.user.findById.mockReturnValueOnce({ user_privileges: ['ADMIN'] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['NO_PRIVILEGES'] });
     db.submission.findById.mockReturnValueOnce({ msg: 'test submission' });
     db.submission.withdrawSubmission.mockReturnValue({ msg: 'submission withdrawn' });
@@ -232,7 +232,7 @@ describe('submission', () => {
       },
       id: 'test id'
     };
-    db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_ADMINREAD'] });
+    db.user.findById.mockReturnValueOnce({ user_privileges: ['ADMIN'] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['NO_PRIVILEGES'] });
     db.submission.findById.mockReturnValueOnce({ msg: 'test submission' });
     db.submission.restoreSubmission.mockReturnValue({ msg: 'submission restored' });
@@ -249,7 +249,7 @@ describe('submission', () => {
       id: 'test id',
       step_name: 'test step'
     };
-    db.user.findById.mockReturnValueOnce({ user_privileges: ['REQUEST_ADMINREAD'] });
+    db.user.findById.mockReturnValueOnce({ user_privileges: ['ADMIN'] });
     db.user.findById.mockReturnValueOnce({ user_privileges: ['NO_PRIVILEGES'] });
     db.submission.setStep.mockReturnValueOnce({ msg: 'step changed' });
     db.submission.findById.mockReturnValueOnce({ msg: 'test submission' });